moving is_active_subscription function to utils
