+++ corehq/apps/accounting/utils.py
@@ -97,3 +97,8 @@ class LazyEncoder(json.JSONEncoder):
         if isinstance(obj, Promise):
             return force_unicode(obj)
         return super(LazyEncoder, self).default(obj)
+
+
+def is_active_subscription(date_start, date_end):
+    today = datetime.date.today()
+    return (date_start is None or date_start <= today) and (date_end is None or today <= date_end)
